requirements.txt
pandas
transformers
bitsandbytes
accelerate
llama-index-core
llama-index-llms-huggingface
fastapi
pydantic
uvicorn
